requirements.txt
dj-database-url==0.5.0
Django==4.0.2
gunicorn==20.1.0
psycopg2-binary==2.9.3
pusher==3.2.0
python-dotenv==0.19.2
whitenoise==6.0.0